Added config files to repository.
Previously, only template files were stored here because
the SMTP credentials were stored in plaintext. Instead, the
credentials are stored as environment variables on the host server and
the corresponding sections were removed from the config file. When
the config file is loaded, the credentials are pulled from the envvars.