Updated supervisor to load /etc/environment
Supervisor subprocesses run in a separate shell from supervisor. They do
not inherit the environment variables of the shell.
`bin/run_with_env.sh` will `source` a file prior to running a command,
such that the environment variables are defined for that command.

For our use case, this passes the AWS SMTP environment variables to the
`runqueue.py` process.